infrastructure/terraform/files/temp_db_init/requirements.txt
psycopg2-binary>=2.9.5
boto3>=1.26.0
